requirements.txt
# To install the dependencies, run the following command:
# pip install -r requirements.txt

# Library for the SQL connection
mysql-connector-python


# Library for the management of .env
python-dotenv

# Library for the management of the API
fastapi